Move picamera threads
